requirements.txt
requests
igdb-api-v4
python-dotenv
sqlmodel
SQLAlchemy-Utils
